--- a/src/fonts.jl
+++ b/src/fonts.jl
@@ -1,3 +1,7 @@
+################################################################################
+#                            General font handling                             #
+################################################################################
+
 function set_font_matrix(cr, matrix)
     ccall((:cairo_set_font_matrix, LIB_CAIRO), Cvoid, (Ptr{Cvoid}, Ptr{Cvoid}), cr.ptr, Ref(matrix))
 end
@@ -34,4 +38,58 @@ function best_font(c::Char, font = AbstractPlotting.defaultfont())
         return AbstractPlotting.defaultfont()
     end
     return font
+end
+
+################################################################################
+#                                Glyph handling                                #
+################################################################################
+
+function glyph_index(char::Char, font::AbstractPlotting.FreeTypeAbstraction.FTFont)
+    return AbstractPlotting.FreeType.FT_Get_Char_Index(font, Culong(char))
+end
+
+"""
+    CairoGlyph(index, x, y)
+    CairoGlyph(char, font, [x = 0.0, y = 0.0])
+
+Constructs a glyph type for Cairo, which stores an index and an offset.
+"""
+struct CairoGlyph
+    "The index of the character in its font"
+    index::Culong
+    """
+    The offset in the X direction between the origin used for drawing or
+    measuring the string and the origin of this glyph.
+    """
+    x::Cdouble
+    """
+    the offset in the Y direction between the origin used for drawing or
+    measuring the string and the origin of this glyph.
+    """
+    y::Cdouble
+end
+
+function CairoGlyph(
+        char::Char, font::AbstractPlotting.FreeTypeAbstraction.FTFont,
+        x::Float64 = 0.0, y::Float64 = 0.0
+    )
+    return CairoGlyph(glyph_index(char, font), x, y)
+end
+
+function show_glyphs(ctx::Cairo.CairoContext, glyphs::Vector{CairoGlyph})
+    return ccall(
+        (:cairo_show_glyphs, LIB_CAIRO),
+        Cvoid,
+        (Ptr{Cvoid}, Ptr{CairoGlyph}, Cint),
+        ctx.ptr, glyphs, length(glyphs)
+    )
+end
+
+function glyph_path(ctx::Cairo.CairoContext, glyphs::Vector{CairoGlyph})
+    return ccall(
+        (:cairo_glyph_path, LIB_CAIRO),
+        Cvoid,
+        (Ptr{Cvoid}, Ptr{CairoGlyph}, Cint),
+        ctx.ptr, glyphs, length(glyphs)
+    )
 end