This is still early stage, and has yet to incorporate marker batching.

## Benchmarks:

### With glyphs
```julia
julia> @benchmark save("streamplot.png", scene)
BenchmarkTools.Trial: 
  memory estimate:  8.56 MiB
  allocs estimate:  359714
  --------------
  minimum time:     125.084 ms (0.00% GC)
  median time:      134.300 ms (0.00% GC)
  mean time:        138.612 ms (1.79% GC)
  maximum time:     174.034 ms (13.50% GC)
  --------------
  samples:          37
  evals/sample:     1

julia> @benchmark save("streamplot.pdf", scene)
BenchmarkTools.Trial: 
  memory estimate:  18.40 MiB
  allocs estimate:  682261
  --------------
  minimum time:     524.890 ms (0.00% GC)
  median time:      567.912 ms (0.00% GC)
  mean time:        567.293 ms (0.61% GC)
  maximum time:     595.558 ms (2.69% GC)
  --------------
  samples:          9
  evals/sample:     1

julia> @benchmark save("streamplot.svg", scene)
BenchmarkTools.Trial: 
  memory estimate:  8.55 MiB
  allocs estimate:  359572
  --------------
  minimum time:     128.375 ms (0.00% GC)
  median time:      136.864 ms (0.00% GC)
  mean time:        141.726 ms (1.91% GC)
  maximum time:     178.830 ms (14.72% GC)
  --------------
  samples:          36
  evals/sample:     1
```
### Without glyphs
```julia
julia> @benchmark save("streamplot.png", scene)
BenchmarkTools.Trial: 
  memory estimate:  8.54 MiB
  allocs estimate:  359714
  --------------
  minimum time:     123.997 ms (0.00% GC)
  median time:      134.350 ms (0.00% GC)
  mean time:        139.012 ms (1.95% GC)
  maximum time:     177.985 ms (15.72% GC)
  --------------
  samples:          37
  evals/sample:     1

julia> @benchmark save("streamplot.pdf", scene)
BenchmarkTools.Trial: 
  memory estimate:  18.39 MiB
  allocs estimate:  682261
  --------------
  minimum time:     534.316 ms (0.00% GC)
  median time:      549.414 ms (0.00% GC)
  mean time:        548.996 ms (0.55% GC)
  maximum time:     565.823 ms (0.00% GC)
  --------------
  samples:          10
  evals/sample:     1

julia> @benchmark save("streamplot.svg", scene)
BenchmarkTools.Trial: 
  memory estimate:  8.54 MiB
  allocs estimate:  359572
  --------------
  minimum time:     131.421 ms (0.00% GC)
  median time:      143.389 ms (0.00% GC)
  mean time:        145.518 ms (2.11% GC)
  maximum time:     180.238 ms (15.33% GC)
  --------------
  samples:          35
  evals/sample:     1
```

These benchmarks were taken from the streamplot example.