@SimonDanisch 2020-10-23
If we switch to glyphs, we'll may reopen this PR.